requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.1
PyPDF2>=3.0.1
python-docx>=1.1.0
scikit-learn>=1.5.0
nltk>=3.8
